Wip invisible (#26)
* Add support for invisible recaptcha and when prototypejs is not yet loaded

* WIP - refacyoring for invisible reCaptcha

* working? frontend with invisible.

* working admin invisible

* Update system.xml

* Add position setting for invisible. Cleanup of admin settings dependency display of depended values and options

* update to wok with checkout

* update to deal with checkout as guest, and register

* add newsletter subscribe recapctha

* add newsletter subscribe recapctha

* drop skin image which was for v1 and is no longer required

* set default theme to invisible

* solve found issue with error: grecaptcha.execute only takes the 'action' parameter.

* Remove MAP in composer. 

There is a modman file that controls deploy, this is duplication

* adjust to solve js error when subscribe captcha is cached and multiple captchas are rendered on one page

* add uninstall potential issue solving

* refine readme

* refine readme

* adjustments to deal with multiple page recaptchas

* Update README.md

* Prevent issue with multiple recapctha elements sharing same ID

It was found that if a page has two of the same captcha areas (example two contact us forms), then the captcha IDs are identical. Add a randomness factor to the id generation.

* Add callback response flags

Add in recaptchaPassBackRecaptchaResponse which allows passing back the response to parentMethod, if set.
Add in recaptchaHasInterceptedFormSubmit flag which allows form submit event listener from not actioning form submit. Useful if form that uses recaptcha is done by ajax form submit event.

* Remove logged in recaptcha limitation

It has been found that bot can create accounts, or use a human created account, and then push in bulk spam newsetter subscriptions.
The code is adjusted to always use recaptcha, and not hide if user is logged in.

With invisible recapctha this is no longer an issue anyways

* update was done in wrong branch. Move code to wip-invisible

* Update recaptcha.phtml

* Merge invisible to become defacto default (#25)

* Update frontend recaptcha template to accomodate ajax submits from recaptcha forms

* restore actual master branch code