Fix backend CAPTCHA form using invalid class
On line 22 the file tries to declare a `new VarienForm`, but this fails because this class does not exist in the backend.  Mage backend uses `js/mage/adminhtml/form.js` for form validation, NOT `js/varien/form.js`.  The latter declares `VarienForm` but the former declares `varienForm`.  Therefore in the backend class `varienForm` must be used.  

Additionally line 22 tries to initialise a variable *called* `varienForm`, which is already declared as a class by `js/mage/adminhtml/form.js`.  Therefore I changed this variable to `adminVarienForm` here, and on line 60 where it is used.

On line 76 the correct declared `varienForm` is used, rather than `VarienForm` which is undefined in the backend.

The above errors stop admin login with reCaptcha enabled from functioning.

Lastly, on line 4, some inline CSS is added to prevent a rogue bullet appearing on the admin form.